+++ SocialMedia/src/main/java/com/SocialMedia/SocialMedia/Service/RequestServiceImpl.java
@@ -5,6 +5,7 @@ import com.SocialMedia.SocialMedia.DTO.RequestDTO;
 import com.SocialMedia.SocialMedia.Entities.Friend;
 import com.SocialMedia.SocialMedia.Entities.Request;
 import com.SocialMedia.SocialMedia.Entities.User;
+import com.SocialMedia.SocialMedia.Exceptions.EntityAlreadyExistsException;
 import com.SocialMedia.SocialMedia.Exceptions.EntityNotFoundException;
 import com.SocialMedia.SocialMedia.Repository.FriendRepo;
 import com.SocialMedia.SocialMedia.Repository.RequestRepo;
@@ -14,6 +15,7 @@ import com.SocialMedia.SocialMedia.Util.AuthenticatedUserUtil;
 import com.SocialMedia.SocialMedia.Util.PaginatedResult;
 import com.amazonaws.services.dynamodbv2.datamodeling.QueryResultPage;
 import com.amazonaws.services.dynamodbv2.model.AttributeValue;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -37,21 +39,46 @@ public class RequestServiceImpl implements RequestService{
     @Autowired
     private FriendRepo friendRepo;
 
+    @Autowired
+    private ModelMapper modelMapper;
+
     @Override
-    public Request sendRequest(RequestDTO requestDTO) throws EntityNotFoundException {
+    public RequestDTO sendRequest(RequestDTO requestDTO) throws EntityNotFoundException {
+        String userId= authUtil.getCurrentUser();
+
+        Request request= modelMapper.map(requestDTO,Request.class);
+
         User user = userRepo.getUserByUsername(requestDTO.getReceiverId());
         if(user==null) throw  new EntityNotFoundException("User Not Found");
-        requestDTO.setSenderId(authUtil.getCurrentUser());
+
+        Request request1= requestRepo.getSentRequest(requestDTO.getReceiverId());
+
+        if(friendRepo.getFriend(userId, requestDTO.getReceiverId())!=null){
+            throw new EntityAlreadyExistsException("You both are friends already");
+        }
+
+        if(request!=null){
+            throw new EntityAlreadyExistsException("Already sent Request");
+        }
+
+        requestDTO.setSenderId(userId);
         requestDTO.setStatus(Status.Pending);
         requestDTO.setSentTime(LocalDateTime.now());
-        return requestRepo.createRequest(requestDTO);
+        requestRepo.createRequest(request);
+        return requestDTO;
     }
 
     @Override
-    public Request withdrawRequest(String receiverId){
+    public RequestDTO withdrawRequest(String receiverId) throws EntityNotFoundException {
 
         String senderId= authUtil.getCurrentUser();
-        return requestRepo.deleteRequest(senderId, receiverId);
+        Request request= requestRepo.getSentRequest(receiverId);
+
+        if(request==null) throw new EntityNotFoundException("Request not found");
+        requestRepo.deleteRequest(senderId, receiverId);
+        RequestDTO requestDTO= modelMapper.map(request,RequestDTO.class);
+        return requestDTO;
+
     }
 
     @Override
@@ -80,25 +107,40 @@ public class RequestServiceImpl implements RequestService{
 
 
     @Override
-    public Request resolveRequest(String senderId, Status status) {
+    public RequestDTO resolveRequest(String senderId, Status status) throws EntityNotFoundException {
         String receiverId = authUtil.getCurrentUser();
         Request request = requestRepo.getreceivedRequest(senderId);
 
         if (request == null) {
-            throw new RuntimeException("Request not found");
+            throw new EntityNotFoundException("Request not found");
         }
 
         if (status == Status.Accept) {
-            FriendDTO friendDTO= new FriendDTO(senderId,receiverId,LocalDateTime.now());
-            FriendDTO friendDTO1= new FriendDTO(receiverId,senderId,LocalDateTime.now());
-            Friend friend = friendRepo.addFriend(friendDTO);
-            Friend friend1= friendRepo.addFriend(friendDTO1);
-            requestRepo.deleteRequest(senderId,receiverId); // Remove request after acceptance
+
+            FriendDTO friendDTO1 = new FriendDTO(senderId, receiverId, LocalDateTime.now());
+            Friend friend1= modelMapper.map(friendDTO1, Friend.class);
+            FriendDTO friendDTO2 = new FriendDTO(receiverId, senderId, LocalDateTime.now());
+            Friend friend2= modelMapper.map(friendDTO2, Friend.class);
+
+            friendRepo.addFriend(friend1);
+            friendRepo.addFriend(friend2);
+
+
+            requestRepo.deleteRequest(senderId, receiverId);
+
+
+            Request reciprocalRequest = requestRepo.getSentRequest(senderId);
+            if (reciprocalRequest != null) {
+                requestRepo.deleteRequest(receiverId, senderId);
+            }
+
         } else if (status == Status.Reject) {
             request.setStatus(Status.Reject);
-            requestRepo.deleteRequest(senderId,receiverId);
+            requestRepo.deleteRequest(senderId, receiverId);
         }
-        return request;
+
+        RequestDTO requestDTO= modelMapper.map(request,RequestDTO.class);
+        return requestDTO;
     }
 }
